feat: refactor access management by renaming Access to AppAccess and updating related functions
